PLANET-7630: Shared links and meta tags don't show all special characters properly
- Shared links and meta tags don't show all special characters properly
